refactor(core): split insight and grounding workflows
