+++ packages/core/src/ai-model/workflows/grounding/index.ts
@@ -1,0 +1,3 @@
+export { AiLocateElement } from './locate';
+export { AiLocateSection } from './section-locate';
+export { buildSearchAreaConfig } from './search-area';
