+++ packages/core/src/ai-model/workflows/grounding/search-area.ts
@@ -1,0 +1,35 @@
+import type { Rect, UIContext } from '@/types';
+import { cropByRect, scaleImage } from '@midscene/shared/img';
+import { expandSearchArea } from '../../../common';
+import type { SearchAreaConfig } from './types';
+
+export async function buildSearchAreaConfig(options: {
+  context: UIContext;
+  baseRect: Rect;
+}): Promise<SearchAreaConfig> {
+  const { context, baseRect } = options;
+  const scaleRatio = 2;
+  const sectionRect = expandSearchArea(baseRect, context.shotSize);
+
+  const croppedResult = await cropByRect(
+    context.screenshot.base64,
+    sectionRect,
+  );
+
+  const scaledResult = await scaleImage(croppedResult.imageBase64, scaleRatio);
+  return {
+    sourceRect: sectionRect,
+    image: {
+      imageBase64: scaledResult.imageBase64,
+      width: scaledResult.width,
+      height: scaledResult.height,
+    },
+    mapping: {
+      offset: {
+        x: sectionRect.left,
+        y: sectionRect.top,
+      },
+      scale: scaleRatio,
+    },
+  };
+}
