--- packages/core/src/ai-model/workflows/inspect/index.ts
@@ -1,7 +1,0 @@
-export {
-  AiExtractElementInfo,
-  AiJudgeOrderSensitive,
-  AiLocateElement,
-  AiLocateSection,
-  buildSearchAreaConfig,
-} from './inspect';
