+++ packages/core/tests/unit-test/extraction.test.ts
@@ -1,4 +1,4 @@
-import { parseXMLExtractionResponse } from '@/ai-model/workflows/inspect/extraction-parser';
+import { parseXMLExtractionResponse } from '@/ai-model/workflows/insight/extraction-parser';
 import { describe, expect, it } from 'vitest';
 
 describe('parseXMLExtractionResponse', () => {
