+++ packages/core/tests/unit-test/grounding.test.ts
@@ -1,6 +1,6 @@
 import { createLocateResultAdapter } from '@/ai-model/shared/model-locate-result';
-import { pixelBboxToRect } from '@/ai-model/workflows/inspect/locate-result-rect';
-import { mapSearchAreaPixelBboxToOriginalPixelBbox } from '@/ai-model/workflows/inspect/search-area-mapping';
+import { pixelBboxToRect } from '@/ai-model/workflows/grounding/locate-result-rect';
+import { mapSearchAreaPixelBboxToOriginalPixelBbox } from '@/ai-model/workflows/grounding/search-area-mapping';
 import { describe, expect, it } from 'vitest';
 
 const actualPixelBboxAdapter = createLocateResultAdapter({
