+++ packages/core/tests/unit-test/image-preprocess.test.ts
@@ -1,5 +1,5 @@
 import { prepareModelImage } from '@/ai-model/model-adapter/image-preprocess';
-import { buildSearchAreaConfig } from '@/ai-model/workflows/inspect';
+import { buildSearchAreaConfig } from '@/ai-model/workflows/grounding';
 import {
   cropByRect,
   paddingToMatchBlockByBase64,
