+++ packages/core/tests/unit-test/insight-multi-frame.test.ts
@@ -19,9 +19,9 @@ vi.mock('@/ai-model/service-caller/index', async () => {
 });
 
 import { callAI } from '@/ai-model/service-caller/index';
-import { AiExtractElementInfo } from '@/ai-model/workflows/inspect';
+import { AiExtractElementInfo } from '@/ai-model/workflows/insight';
 
-describe('AiExtractElementInfo multi-frame context', () => {
+describe('insight extraction multi-frame context', () => {
   const modelConfig: IModelConfig = {
     modelFamily: 'qwen2.5-vl',
     modelName: 'test-model',
