+++ packages/core/tests/unit-test/model-adapter/auto-glm/locate.test.ts
@@ -2,8 +2,8 @@ import { ResolvedModelAdapter } from '@/ai-model/model-adapter/resolve';
 import { autoGlmAdapters } from '@/ai-model/models/auto-glm/adapter';
 import { createAutoGlmPlanningTapLocator } from '@/ai-model/models/auto-glm/locate';
 import { callAIWithStringResponse } from '@/ai-model/service-caller/index';
-import { AiLocateElement } from '@/ai-model/workflows/inspect';
-import type { LocateOptions } from '@/ai-model/workflows/inspect/types';
+import { AiLocateElement } from '@/ai-model/workflows/grounding';
+import type { LocateOptions } from '@/ai-model/workflows/grounding/types';
 import type { UIContext } from '@/types';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
